fix(naming): Rename all SolarEvent related classes and methods to avoid confusion on them being plural
