+++ src/test/java/com/codecool/solarwatch/service/SolarEventServiceTest.java
@@ -7,7 +7,7 @@ import com.codecool.solarwatch.client.response.SolarEventsResponse;
 import com.codecool.solarwatch.exception.ResourceNotFoundException;
 import com.codecool.solarwatch.model.City;
 import com.codecool.solarwatch.model.SolarEvent;
-import com.codecool.solarwatch.model.dto.SolarEventsDTO;
+import com.codecool.solarwatch.model.dto.SolarEventDTO;
 import com.codecool.solarwatch.repository.CityRepository;
 import com.codecool.solarwatch.repository.SolarEventRepository;
 import org.junit.jupiter.api.DisplayName;
@@ -27,7 +27,7 @@ import static org.mockito.Mockito.*;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-class SolarEventsServiceTest {
+class SolarEventServiceTest {
 
   @Mock private GeoClient geoClient;
   @Mock private SolarEventsClient solarClient;
@@ -35,7 +35,7 @@ class SolarEventsServiceTest {
   @Mock private SolarEventRepository solarEventRepository;
 
   @InjectMocks
-  private SolarEventsService service;
+  private SolarEventService service;
 
   private final LocalDate DATE = LocalDate.of(2023, 10, 5);
   private final String CITY_NAME = "London";
@@ -43,14 +43,14 @@ class SolarEventsServiceTest {
 
   @Test
   @DisplayName("Should return cached event when City and Event exist")
-  void getSolarEvents_FullCacheHit() {
+  void getSolarEvent_FullCacheHit() {
     City city = new City(1L, CITY_NAME, 51.5, -0.1, "GB", "England", null);
     SolarEvent event = new SolarEvent(10L, city, DATE, "6:00 AM", "6:00 PM", "UTC");
 
     when(cityRepository.findByNameIgnoreCase(CITY_NAME)).thenReturn(Optional.of(city));
     when(solarEventRepository.findByCityAndDate(city, DATE)).thenReturn(Optional.of(event));
 
-    SolarEventsDTO result = service.getSolarEvents(CITY_NAME, DATE);
+    SolarEventDTO result = service.getSolarEvent(CITY_NAME, DATE);
 
     assertNotNull(result);
     assertEquals(CITY_NAME, result.city());
@@ -62,14 +62,14 @@ class SolarEventsServiceTest {
 
   @Test
   @DisplayName("Should return cached event even when input case differs from DB case")
-  void getSolarEvents_CaseInsensitiveCacheHit() {
+  void getSolarEvent_CaseInsensitiveCacheHit() {
     City city = new City(1L, CITY_NAME, 51.5, -0.1, "GB", "England", null);
     SolarEvent event = new SolarEvent(10L, city, DATE, "7:00 AM", "7:00 PM", "UTC");
 
     when(cityRepository.findByNameIgnoreCase(CITY_NAME_LOWER)).thenReturn(Optional.of(city));
     when(solarEventRepository.findByCityAndDate(city, DATE)).thenReturn(Optional.of(event));
 
-    SolarEventsDTO result = service.getSolarEvents(CITY_NAME_LOWER, DATE);
+    SolarEventDTO result = service.getSolarEvent(CITY_NAME_LOWER, DATE);
 
     assertEquals(CITY_NAME, result.city());
 
@@ -79,7 +79,7 @@ class SolarEventsServiceTest {
 
   @Test
   @DisplayName("Should fetch City and Event from API when cache misses completely")
-  void getSolarEvents_FullCacheMiss() {
+  void getSolarEvent_FullCacheMiss() {
     GeoResponse geoResponse = new GeoResponse(CITY_NAME, 51.5, -0.1, "GB", "England");
     SolarEventsResponse solarResponse = new SolarEventsResponse(
             new SolarEventsResponse.Results("6:00 AM", "8:00 PM"), "UTC");
@@ -95,7 +95,7 @@ class SolarEventsServiceTest {
     when(solarClient.fetchSolarEvents(51.5, -0.1, DATE)).thenReturn(solarResponse);
     when(solarEventRepository.save(any(SolarEvent.class))).thenAnswer(i -> i.getArguments()[0]);
 
-    SolarEventsDTO result = service.getSolarEvents(CITY_NAME, DATE);
+    SolarEventDTO result = service.getSolarEvent(CITY_NAME, DATE);
 
     assertEquals(CITY_NAME, result.city());
     verify(cityRepository, times(2)).findByNameIgnoreCase(CITY_NAME);
@@ -104,7 +104,7 @@ class SolarEventsServiceTest {
 
   @Test
   @DisplayName("Should fetch Event from API when City exists but Event does not")
-  void getSolarEvents_CityHit_EventMiss() {
+  void getSolarEvent_CityHit_EventMiss() {
     City city = new City(1L, CITY_NAME, 51.5, -0.1, "GB", "England", null);
     SolarEventsResponse apiResponse = new SolarEventsResponse(
             new SolarEventsResponse.Results("7:00 AM", "7:00 PM"), "UTC");
@@ -114,7 +114,7 @@ class SolarEventsServiceTest {
     when(solarClient.fetchSolarEvents(city.getLatitude(), city.getLongitude(), DATE)).thenReturn(apiResponse);
     when(solarEventRepository.save(any(SolarEvent.class))).thenAnswer(i -> i.getArguments()[0]);
 
-    SolarEventsDTO result = service.getSolarEvents(CITY_NAME, DATE);
+    SolarEventDTO result = service.getSolarEvent(CITY_NAME, DATE);
 
     assertEquals("7:00 AM", result.sunrise());
     verify(solarEventRepository).save(any(SolarEvent.class));
@@ -122,12 +122,12 @@ class SolarEventsServiceTest {
 
   @Test
   @DisplayName("Should throw ResourceNotFoundException when City not found in DB or API")
-  void getSolarEvents_CityNotFoundAnywhere() {
+  void getSolarEvent_CityNotFoundAnywhere() {
     String unknownCity = "NotARealCity";
     when(cityRepository.findByNameIgnoreCase(unknownCity)).thenReturn(Optional.empty());
     when(geoClient.fetchCityData(eq(unknownCity), anyInt(), any())).thenReturn(List.of());
 
-    assertThrows(ResourceNotFoundException.class, () -> service.getSolarEvents(unknownCity, DATE));
+    assertThrows(ResourceNotFoundException.class, () -> service.getSolarEvent(unknownCity, DATE));
   }
 
 }
